+++ axelor-production/src/main/java/com/axelor/apps/production/web/ManufOrderController.java
@@ -19,6 +19,7 @@ package com.axelor.apps.production.web;
 
 import java.io.IOException;
 import java.lang.invoke.MethodHandles;
+import java.math.BigDecimal;
 import java.util.List;
 
 import org.eclipse.birt.core.exception.BirtException;
@@ -33,7 +34,6 @@ import com.axelor.apps.production.report.IReport;
 import com.axelor.apps.production.service.ManufOrderService;
 import com.axelor.apps.production.service.ManufOrderWorkflowService;
 import com.axelor.apps.report.engine.ReportSettings;
-import com.axelor.apps.stock.db.StockMove;
 import com.axelor.exception.AxelorException;
 import com.axelor.exception.service.TraceBackService;
 import com.axelor.i18n.I18n;
@@ -103,7 +103,7 @@ public class ManufOrderController {
 		
 	}
 	
-	public void finish (ActionRequest request, ActionResponse response) throws AxelorException {
+	public void finish (ActionRequest request, ActionResponse response) {
 		
 		try  {
 			Long manufOrderId = (Long)request.getContext().get("id");
@@ -117,8 +117,20 @@ public class ManufOrderController {
 		}	
 		
 	}
-	
-	public void cancel (ActionRequest request, ActionResponse response) throws AxelorException {
+
+	public void partialFinish(ActionRequest request, ActionResponse response) {
+		try {
+			ManufOrder manufOrder = request.getContext().asType(ManufOrder.class);
+			manufOrder = manufOrderRepo.find(manufOrder.getId());
+
+			Beans.get(ManufOrderWorkflowService.class).partialFinish(manufOrder);
+			response.setReload(true);
+		} catch (Exception e) {
+			TraceBackService.trace(response, e);
+		}
+	}
+
+	public void cancel (ActionRequest request, ActionResponse response) {
 		
 		try  {
 			Long manufOrderId = (Long)request.getContext().get("id");
@@ -132,7 +144,7 @@ public class ManufOrderController {
 		}	
 	}
 	
-	public void plan (ActionRequest request, ActionResponse response) throws AxelorException {
+	public void plan (ActionRequest request, ActionResponse response) {
 		
 		try  {
 			Long manufOrderId = (Long)request.getContext().get("id");
@@ -157,7 +169,7 @@ public class ManufOrderController {
 	 * @throws BirtException 
 	 * @throws IOException 
 	 */
-	public void print(ActionRequest request, ActionResponse response) throws AxelorException {
+	public void print(ActionRequest request, ActionResponse response) {
 		
 		try  {
 			ManufOrder manufOrder = request.getContext().asType( ManufOrder.class );
@@ -221,22 +233,47 @@ public class ManufOrderController {
 		
 	}
 
-	public void generateWasteStockMove(ActionRequest request, ActionResponse response) throws AxelorException {
+	public void generateWasteStockMove(ActionRequest request, ActionResponse response) {
 		try  {
 			ManufOrder manufOrder = request.getContext().asType(ManufOrder.class);
 			manufOrder = manufOrderRepo.find(manufOrder.getId());
-			StockMove wasteStockMove = manufOrderService.generateWasteStockMove(manufOrder);
+			manufOrderService.generateWasteStockMove(manufOrder);
 			response.setReload(true);
 		} catch (Exception e) {
 			TraceBackService.trace(response, e);
 		}
 	}
-  
-	public void updateQty(ActionRequest request, ActionResponse response) {
+
+	/**
+	 * Called from manuf order wizard view.
+	 * Call {@link ManufOrderService#updatePlannedQty(ManufOrder)}
+	 * @param request
+	 * @param response
+	 */
+	public void updatePlannedQty(ActionRequest request, ActionResponse response) {
 		try  {
 			ManufOrder manufOrder = request.getContext().asType(ManufOrder.class);
 			manufOrder = manufOrderRepo.find(manufOrder.getId());
-			manufOrderService.updateQty(manufOrder);
+			manufOrderService.updatePlannedQty(manufOrder);
+			response.setReload(true);
+			response.setCanClose(true);
+		} catch (Exception e) {
+			TraceBackService.trace(response, e);
+		}
+	}
+
+	/**
+	 * Called from manuf order wizard view.
+	 * Call {@link ManufOrderService#updateRealQty(ManufOrder, BigDecimal)}
+	 * @param request
+	 * @param response
+	 */
+	public void updateRealQty(ActionRequest request, ActionResponse response) {
+		try  {
+			ManufOrder manufOrder = request.getContext().asType(ManufOrder.class);
+			manufOrder = manufOrderRepo.find(manufOrder.getId());
+			BigDecimal qtyToUpdate = new BigDecimal(request.getContext().get("qtyToUpdate").toString());
+			manufOrderService.updateRealQty(manufOrder, qtyToUpdate);
 			response.setReload(true);
 			response.setCanClose(true);
 		} catch (Exception e) {
@@ -244,7 +281,7 @@ public class ManufOrderController {
 		}
 	}
 	
-	public void printProdProcess(ActionRequest request, ActionResponse response) throws AxelorException {
+	public void printProdProcess(ActionRequest request, ActionResponse response) {
 		
 		try  {
 			ManufOrder manufOrder = request.getContext().asType( ManufOrder.class );
@@ -266,7 +303,7 @@ public class ManufOrderController {
 		
 	}
 
-	public void updatePlannedDates(ActionRequest request, ActionResponse response) throws AxelorException {
+	public void updatePlannedDates(ActionRequest request, ActionResponse response) {
 		
 		try  {
 			ManufOrder manufOrderView = request.getContext().asType(ManufOrder.class);